+++ packages/analytics-js/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.31.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.30.0...@rudderstack/analytics-js@3.31.0) (2026-03-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.28.1`
+
+### Features
+
+* **analytics-js:** replace error-stack-parser with lean inline parser [SDK-4625] ([#2833](https://github.com/rudderlabs/rudder-sdk-js/issues/2833)) ([9a99057](https://github.com/rudderlabs/rudder-sdk-js/commit/9a990577e928b007c88f665ecec27c1fd6fa8afe))
+
 ## [3.30.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.29.0...@rudderstack/analytics-js@3.30.0) (2026-03-02)