+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.52.11](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.9...rudder-sdk-js@2.52.11) (2026-03-23)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-integrations` updated to version `3.20.1`
 ## [2.52.9](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.52.8...rudder-sdk-js@2.52.9) (2026-03-02)
 
 ### Dependency Updates